test(#1714): split the shebang cases out and share one scratch-crate harness
The 1450-line test file had no headroom under the ~1500 test ratchet, and each
integration test is its own crate — so a second test file either copies the
scratch-crate harness or shares it. Copies drift, so `ScratchCrate` and the
`stripped` sanitizer wrapper move to `support/mod_reachability_harness.rs` and
both files include it.

The shebang cases move to their own `issue_1714_mod_reachability_shebang.rs`:
rustc's shebang rule is narrow and non-obvious, both directions of it are
load-bearing, and keeping the whole table in one file is what lets the next
reader see it at once.

Pure move — no behavior change, 33/33 still green (29 + 4).
