+++ cqlite-core/tests/issue_1714_mod_reachability.rs
@@ -28,13 +28,15 @@
 //! The walker itself lives in `support/mod_reachability.rs` and is crate-agnostic so
 //! #1502 (the `cqlite-cli` mod-wiring guard) can drive it for `src/main.rs`.
 
+#[path = "support/mod_reachability_harness.rs"]
+mod harness;
 #[path = "support/mod_reachability.rs"]
 mod mod_reachability;
 
 use std::fs;
 use std::path::{Path, PathBuf};
-use std::sync::atomic::{AtomicUsize, Ordering};
 
+use harness::{stripped, ScratchCrate};
 use mod_reachability::{analyze, strip_comments_and_strings, ExpectedOrphan, ModuleGraphSpec};
 
 const SRC_DIR: &str = "src";
@@ -286,81 +288,6 @@ pub fn lifetimes<'a>(s: &'a str) -> &'a str { s }
     );
 }
 
-/// rustc DISCARDS a `#!` line at byte offset 0 of a source file, so whatever that line
-/// says is not Rust at all. Parsing it as ordinary code is the false-PASS path this
-/// asserts against: a `#!/usr/bin/env -S tool mod orphan;` first line would otherwise
-/// yield a bogus `mod orphan;` declaration and make an unreachable file look reachable —
-/// the exact silent green this guard exists to prevent (#1714).
-#[test]
-fn shebang_line_at_offset_zero_does_not_declare_a_module() {
-    let tree = ScratchCrate::new("shebang-not-a-decl");
-    tree.write(
-        "src/lib.rs",
-        "#!/usr/bin/env -S tool mod orphan;\npub mod wired;\n",
-    );
-    tree.write("src/wired.rs", "pub fn f() {}\n");
-    tree.write("src/orphan.rs", "pub fn never_compiled() {}\n");
-
-    let report = tree.analyze();
-    assert_eq!(
-        report.orphans.iter().cloned().collect::<Vec<_>>(),
-        vec!["src/orphan.rs".to_string()],
-        "a `mod orphan;` inside a SHEBANG line (which rustc discards) must NOT make \
-         `orphan.rs` reachable; enumerated={:?} reachable={:?}",
-        report.enumerated,
-        report.reachable
-    );
-    assert!(
-        report.reachable.contains("src/wired.rs"),
-        "a shebang-bearing root must still resolve the real declarations below it"
-    );
-}
-
-/// The other direction of the same fix, and the one that would be catastrophic to get
-/// backwards: `#![...]` at byte offset 0 is a Rust INNER ATTRIBUTE (this crate's own
-/// `lib.rs` opens with one), not a shebang. Blanking it would eat the rest of that line,
-/// so the real `pub mod wired;` sharing it must survive.
-#[test]
-fn inner_attribute_at_offset_zero_is_not_mistaken_for_a_shebang() {
-    let tree = ScratchCrate::new("inner-attr-not-shebang");
-    tree.write("src/lib.rs", "#![allow(dead_code)] pub mod wired;\n");
-    tree.write("src/wired.rs", "pub fn f() {}\n");
-
-    let report = tree.analyze();
-    assert!(
-        report.orphans.is_empty(),
-        "`#![allow(dead_code)]` at offset 0 is an inner attribute, not a shebang — \
-         blanking its line swallowed the real `pub mod wired;`: orphans={:?}",
-        report.orphans
-    );
-    let out = stripped("#![allow(dead_code)] pub mod wired;\n");
-    assert_eq!(
-        out, "#![allow(dead_code)] pub mod wired;\n",
-        "an inner attribute at offset 0 must reach the parser byte-for-byte"
-    );
-}
-
-/// The shebang rule applies at BYTE OFFSET 0 and nowhere else: an implementation that
-/// stripped any line merely *starting* with `#!` would eat this file's real declaration,
-/// which sits on the ordinary `#![...]`-after-a-comment-header line every Rust crate has.
-#[test]
-fn hash_bang_after_offset_zero_is_not_a_shebang() {
-    let tree = ScratchCrate::new("hashbang-not-at-zero");
-    tree.write(
-        "src/lib.rs",
-        "//! Crate docs, line 1.\n//! Line 2.\n//! Line 3.\n\n#![allow(dead_code)] pub mod wired;\n",
-    );
-    tree.write("src/wired.rs", "pub fn f() {}\n");
-
-    let report = tree.analyze();
-    assert!(
-        report.orphans.is_empty(),
-        "a `#!` line that is NOT at byte offset 0 is not a shebang; its declarations must \
-         survive: orphans={:?}",
-        report.orphans
-    );
-}
-
 /// Every resolution rule, exercised against a tree the test controls end to end.
 #[test]
 fn synthetic_tree_exercises_every_resolution_rule() {
@@ -564,8 +491,11 @@ fn symlink_under_src_fails_closed() {
     tree.write("src/lib.rs", "pub mod wired;\n");
     tree.write("src/wired.rs", "pub fn f() {}\n");
     tree.write("outside/hidden_orphan.rs", "pub fn never_compiled() {}\n");
-    symlink(tree.dir.join("outside"), tree.dir.join("src/linked_dir"))
-        .unwrap_or_else(|e| panic!("cannot create directory symlink: {e}"));
+    symlink(
+        tree.dir().join("outside"),
+        tree.dir().join("src/linked_dir"),
+    )
+    .unwrap_or_else(|e| panic!("cannot create directory symlink: {e}"));
     let cause = tree.expect_failure();
     assert!(
         cause.contains("symlink") && cause.contains("linked_dir"),
@@ -578,8 +508,8 @@ fn symlink_under_src_fails_closed() {
     tree.write("src/wired.rs", "pub fn f() {}\n");
     tree.write("outside/target.rs", "pub fn f() {}\n");
     symlink(
-        tree.dir.join("outside/target.rs"),
-        tree.dir.join("src/linked.rs"),
+        tree.dir().join("outside/target.rs"),
+        tree.dir().join("src/linked.rs"),
     )
     .unwrap_or_else(|e| panic!("cannot create file symlink: {e}"));
     let cause = tree.expect_failure();
@@ -1241,10 +1171,6 @@ fn the_non_ascii_identifier_refusal_never_reds_ordinary_rust() {
 // Stripper unit tests (control vs data, at the sanitizer boundary)
 // ---------------------------------------------------------------------------
 
-fn stripped(src: &str) -> String {
-    strip_comments_and_strings(src).unwrap_or_else(|e| panic!("sanitize failed: {e}"))
-}
-
 #[test]
 fn stripper_blanks_every_comment_form() {
     for src in [
@@ -1329,122 +1255,5 @@ fn stripper_fails_closed_on_unterminated_input() {
             strip_comments_and_strings(src).is_err(),
             "unterminated {what} must fail closed rather than yield a guessed parse"
         );
-    }
-}
-
-/// A shebang line is blanked like every other non-Rust span: same byte length, newline
-/// preserved, so byte offsets and reported line numbers stay aligned with the original
-/// source. The both-directions half is asserted too — `#![...]` at offset 0 is an inner
-/// attribute and must reach the parser untouched — plus the fail-closed refusal for the
-/// one shape whose shebang-vs-attribute reading this walker does not model.
-#[test]
-fn stripper_blanks_a_shebang_but_never_an_inner_attribute() {
-    let src = "#!/usr/bin/env cargo\nmod a;\n";
-    let out = stripped(src);
-    assert_eq!(
-        out, "                    \nmod a;\n",
-        "the shebang line must be blanked in place (offsets and newline preserved): {out:?}"
-    );
-    assert_eq!(out.len(), src.len(), "byte offsets must be preserved");
-    assert_eq!(
-        out.matches('\n').count(),
-        src.matches('\n').count(),
-        "line breaks must be preserved so reported line numbers stay correct"
-    );
-
-    for attr in [
-        "#![allow(dead_code)]\nmod a;\n",
-        "#![doc = \"not a shebang\"]\nmod a;\n",
-    ] {
-        // The literal's own value is blanked (it is data), so compare the code skeleton.
-        let out = stripped(attr);
-        assert!(
-            out.starts_with("#![") && out.contains("mod a;"),
-            "`#![...]` at offset 0 is an inner attribute, not a shebang: {attr:?} -> {out:?}"
-        );
-    }
-
-    // `#!` + whitespace/comment + `[` is an inner attribute to rustc (it skips both
-    // before deciding) and a shebang to a naive next-byte test. This walker models
-    // neither reading, so it refuses rather than guessing (FAIL-CLOSED).
-    for src in [
-        "#! [allow(dead_code)]\nmod a;\n",
-        "#!\n[allow(dead_code)]\nmod a;\n",
-        "#!/* c */[allow(dead_code)]\nmod a;\n",
-        "#!// c\n[allow(dead_code)]\nmod a;\n",
-    ] {
-        assert!(
-            strip_comments_and_strings(src).is_err(),
-            "the unmodeled shebang-vs-inner-attribute shape must fail closed: {src:?}"
-        );
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Scratch crate helper
-// ---------------------------------------------------------------------------
-
-static SCRATCH_SEQ: AtomicUsize = AtomicUsize::new(0);
-
-/// A throwaway crate-shaped directory tree under the OS temp dir.
-struct ScratchCrate {
-    dir: PathBuf,
-}
-
-impl ScratchCrate {
-    fn new(label: &str) -> Self {
-        let seq = SCRATCH_SEQ.fetch_add(1, Ordering::SeqCst);
-        let dir = std::env::temp_dir().join(format!(
-            "cqlite-1714-mod-reach-{}-{seq}-{label}",
-            std::process::id()
-        ));
-        let _ = fs::remove_dir_all(&dir);
-        fs::create_dir_all(&dir).unwrap_or_else(|e| panic!("cannot create {}: {e}", dir.display()));
-        Self { dir }
-    }
-
-    fn write(&self, rel: &str, contents: &str) {
-        let path = self.dir.join(rel);
-        if let Some(parent) = path.parent() {
-            fs::create_dir_all(parent)
-                .unwrap_or_else(|e| panic!("cannot create {}: {e}", parent.display()));
-        }
-        fs::write(&path, contents)
-            .unwrap_or_else(|e| panic!("cannot write {}: {e}", path.display()));
-    }
-
-    fn spec(&self) -> ModuleGraphSpec {
-        ModuleGraphSpec {
-            crate_dir: self.dir.clone(),
-            root_file_rel: "src/lib.rs".to_string(),
-            src_dir_rel: SRC_DIR.to_string(),
-        }
-    }
-
-    fn analyze(&self) -> mod_reachability::Report {
-        analyze(&self.spec()).unwrap_or_else(|cause| {
-            panic!(
-                "walk of scratch crate {} failed: {cause}",
-                self.dir.display()
-            )
-        })
-    }
-
-    /// Assert the walk fails closed and return the cause.
-    fn expect_failure(&self) -> String {
-        match analyze(&self.spec()) {
-            Ok(report) => panic!(
-                "expected a FAIL-CLOSED refusal, got a report (orphans={:?}) — a skip-and-continue \
-                 here IS the vacuous pass",
-                report.orphans
-            ),
-            Err(cause) => cause,
-        }
-    }
-}
-
-impl Drop for ScratchCrate {
-    fn drop(&mut self) {
-        let _ = fs::remove_dir_all(&self.dir);
     }
 }
